Links to stories with notifications open new tab
Rather than using a (click) directive and
function for routing to the story the
tag is now an <a> tag with target="_blank" (open in a new tab)
and routerLink="{{'/dashboard/' + story._id}}".

Pro:
* There is no chance of someone losing
their work by clicking one of these links because
their tab with potentially unsaved work is left
open.
* The code is simpler

Cons:
* More tabs being open might be confusing.
* It opens in a new tab no matter what,
even if they are already looking at that story.